refactor: UserDao#insert 전략 패턴으로 변경
