+++ app/src/main/java/com/techcourse/dao/InsertStatement.java
@@ -1,0 +1,23 @@
+package com.techcourse.dao;
+
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public class InsertStatement implements StatementStrategy{
+    private User user;
+
+    public InsertStatement(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public PreparedStatement makePreparedStatement(Connection conn) throws SQLException {
+        PreparedStatement pstmt = conn.prepareStatement("insert into users (account, password, email) values (?, ?, ?)");
+        pstmt.setString(1, user.getAccount());
+        pstmt.setString(2, user.getPassword());
+        pstmt.setString(3, user.getEmail());
+        return pstmt;
+    }
+}
